server/requirements.txt
fastapi
uvicorn
pillow
python-dotenv
openai==0.28
python-multipart
redis
ImageHash

